+++ TestAgenda/src/main/java/com/example/TestAgenda/Service/AgendaService.java
@@ -1,6 +1,7 @@
 package com.example.TestAgenda.Services;
 
 import com.example.TestAgenda.Models.Agenda;
+import com.example.TestAgenda.Models.Status;
 import com.example.TestAgenda.Repository.AgendaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,23 +15,37 @@ public class AgendaService {
     @Autowired
     private AgendaRepository agendaRepository;
 
-    // Método para listar todas as agendas
+
     public List<Agenda> findAll() {
         return agendaRepository.findAll();
     }
 
-    // Método para buscar uma agenda por ID
     public Optional<Agenda> findOne(Long id) {
         return agendaRepository.findById(id);
     }
 
-    // Método para salvar ou atualizar uma agenda
     public Agenda save(Agenda agenda) {
+        if (agenda.getId() != null) {
+            Optional<Agenda> agendaExistente = agendaRepository.findById(agenda.getId());
+
+            if (agendaExistente.isPresent()) {
+                if (agendaExistente.get().getStatus() == Status.FINALIZADO) {
+                    throw new IllegalStateException("Não é permitido excluir ou editar uma agenda com status FINALIZADO.");
+                }
+            }
+        }
         return agendaRepository.save(agenda);
     }
 
-    // Método para deletar uma agenda por ID
+
     public void delete(Long id) {
+        Optional<Agenda> agenda = agendaRepository.findById(id);
+
+        if (agenda.isPresent() && agenda.get().getStatus() == Status.FINALIZADO) {
+            throw new IllegalStateException("Não é permitido excluir ou editar uma agenda com status FINALIZADO.");
+        }
+
         agendaRepository.deleteById(id);
     }
+
 }
